requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
--find-links https://download.pytorch.org/whl/torch_stable.html

certifi==2020.11.8
    # via matplotlib
click==7.1.2
    # via pip-tools
cycler==0.10.0
    # via matplotlib
einops==0.3.0
    # via -r requirements.in
importlib-metadata==2.0.0
    # via tikzplotlib
kiwisolver==1.3.1
    # via matplotlib
matplotlib==3.3.2
    # via
    #   -r requirements.in
    #   tikzplotlib
numpy==1.19.4
    # via
    #   -r requirements.in
    #   matplotlib
    #   opencv-python
    #   pandas
    #   scipy
    #   tikzplotlib
    #   torch
opencv-python==4.4.0.46
    # via -r requirements.in
pandas==1.2.1
    # via -r requirements.in
pillow==8.0.1
    # via
    #   -r requirements.in
    #   matplotlib
    #   tikzplotlib
pip-tools==5.5
    # via -r requirements.in
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.1
    # via
    #   matplotlib
    #   pandas
pytz==2021.1
    # via pandas
pyyaml==5.4.1
    # via -r requirements.in
scipy==1.5.4
    # via -r requirements.in
six==1.15.0
    # via
    #   cycler
    #   python-dateutil
tabulate==0.8.9
    # via -r requirements.in
tikzplotlib==0.9.4
    # via -r requirements.in
torch==1.8.1+cu111
    # via -r requirements.in
typing-extensions==3.7.4.3
    # via torch
zipp==3.4.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
